small insufficient changes in utils.feature_preprocessing.py
